Project: move Schema to main files
